% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EK}
\alias{EK}
\title{Eventyrkode (fairy tale code)}
\format{A data frame with nine variables:
\describe{
\item{\code{EK}}{Numerical value for identification of fairy tales}
\item{\code{date}}{Date of publication}
\item{\code{approximate}}{Is the \code{date} an approximation}
\item{\code{org_language}}{Original language of the fairy tale}
\item{\code{name_da}}{Name of the fairy tale in Danish}
\item{\code{name_de}}{Name of the fairy tale in German}
\item{\code{name_en}}{Name of the fairy tale in English}
\item{\code{name_es}}{Name of the fairy tale in Spanish}
\item{\code{name_fr}}{Name of the fairy tale in French}
}}
\usage{
EK
}
\description{
Eventyrkode (fairy tale code), stipulated by Svend Juel Møller in order
to identify the fairy tales by H.C. Andersen
}
\keyword{datasets}
