% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hcandersen_da}
\alias{hcandersen_da}
\title{The text of H.C. Andersens fairy tales in Danish.}
\format{A data frame with 20439 rows and 2 variables: \code{text} and
\code{book}.}
\usage{
hcandersen_da
}
\description{
A dataset containing 139 out of 157 of H.C. andersens fairy tales
translated in Danish. The UTF-8 plain text was sourced from
http://www.andersenstories.com/ and is divided into elements of up
to about 80 characters each.
}
\keyword{datasets}
