% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hcandersen_de}
\alias{hcandersen_de}
\title{The text of H.C. Andersens fairy tales in German.}
\format{A data frame with 28255 rows and 2 variables: \code{text} and
\code{book}.}
\usage{
hcandersen_de
}
\description{
A dataset containing 156 out of 157 of H.C. andersens fairy tales
translated to German. The UTF-8 plain text was sourced from
http://www.andersenstories.com/ and is divided into elements of up
to about 80 characters each.
}
\keyword{datasets}
