% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hcandersen_en}
\alias{hcandersen_en}
\title{The text of H.C. Andersens fairy tales in English.}
\format{A data frame with 27859 rows and 2 variables: \code{text} and
\code{book}.}
\usage{
hcandersen_en
}
\description{
A dataset containing 157 out of 157 of H.C. andersens fairy tales
translated to English. The UTF-8 plain text was sourced from
http://www.andersenstories.com/ and is divided into elements of up
to about 80 characters each.
}
\keyword{datasets}
