% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{hcandersenr}
\alias{hcandersenr}
\title{Fairy tales and stories of H.C. Andersen.}
\description{
This package contains the (mostly) complete text of 157 fairy tales of
H.C. Andersen, in a number of different languages. formatted to be
convenient for text analysis.
}
